Almost completed. Created validation

+++ Academy2018_.NET_Homework4.Core/Services/AirplaneTypesService.cs
@@ -1,11 +1,12 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using Academy2018_.NET_Homework4.Core.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Models;
 using Academy2018_.NET_Homework4.Shared.DTOs;
+using Academy2018_.NET_Homework4.Shared.Exceptions;
 using AutoMapper;
+using FluentValidation;
 
 namespace Academy2018_.NET_Homework4.Core.Services
 {
@@ -13,11 +14,16 @@ namespace Academy2018_.NET_Homework4.Core.Services
     {
         private readonly IRepository<AirplaneType> _repository;
         private readonly IMapper _mapper;
+        private readonly AbstractValidator<AirplaneTypeDto> _validator;
 
-        public AirplaneTypesService(IRepository<AirplaneType> repository, IMapper mapper)
+        public AirplaneTypesService(
+            IRepository<AirplaneType> repository,
+            IMapper mapper,
+            AbstractValidator<AirplaneTypeDto> validator)
         {
             _repository = repository;
             _mapper = mapper;
+            _validator = validator;
         }
 
         public IEnumerable<AirplaneTypeDto> GetAll()
@@ -28,25 +34,62 @@ namespace Academy2018_.NET_Homework4.Core.Services
 
         public AirplaneTypeDto GetById(object id)
         {
-            return _mapper.Map<AirplaneType, AirplaneTypeDto>(
+            var response = _mapper.Map<AirplaneType, AirplaneTypeDto>(
                 _repository.Get().FirstOrDefault(a => a.Id == (int)id));
+
+            if (response == null)
+            {
+                throw new NotExistException();
+            }
+
+            return response;
         }
 
-        public void Add(AirplaneTypeDto dto)
+        public object Add(AirplaneTypeDto dto)
         {
-            _repository.Create(
-                _mapper.Map<AirplaneTypeDto, AirplaneType>(dto));
+            var validationResult = _validator.Validate(dto);
+
+            if (validationResult.IsValid)
+            {
+                return _repository.Create(
+                    _mapper.Map<AirplaneTypeDto, AirplaneType>(dto));
+            }
+
+            throw new ValidationException(validationResult.Errors);
         }
 
         public void Update(object id, AirplaneTypeDto dto)
         {
-            _repository.Update((int)id,
-                _mapper.Map<AirplaneTypeDto, AirplaneType>(dto));
+            if (_repository.IsExist(id))
+            {
+                var validationResult = _validator.Validate(dto);
+
+                if (validationResult.IsValid)
+                {
+                    _repository.Update((int)id,
+                        _mapper.Map<AirplaneTypeDto, AirplaneType>(dto));
+                }
+                else
+                {
+                    throw new ValidationException(validationResult.Errors);
+                }
+            }
+            else
+            {
+                throw new NotExistException();
+            }        
         }
 
         public void Delete(object id)
         {
-            _repository.Delete((int)id);
+            if (_repository.IsExist(id))
+            {
+                _repository.Delete((int)id);
+            }
+            else
+            {
+                throw new NotExistException();
+            }
         }
     }
 }
