+++ Academy2018_.NET_Homework4.Core/Services/DeparturesService.cs
@@ -4,7 +4,9 @@ using Academy2018_.NET_Homework4.Core.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Models;
 using Academy2018_.NET_Homework4.Shared.DTOs;
+using Academy2018_.NET_Homework4.Shared.Exceptions;
 using AutoMapper;
+using FluentValidation;
 
 namespace Academy2018_.NET_Homework4.Core.Services
 {
@@ -12,11 +14,16 @@ namespace Academy2018_.NET_Homework4.Core.Services
     {
         private readonly IRepository<Departure> _repository;
         private readonly IMapper _mapper;
+        private readonly AbstractValidator<DepartureDto> _validator;
 
-        public DeparturesService(IRepository<Departure> repository, IMapper mapper)
+        public DeparturesService(
+            IRepository<Departure> repository, 
+            IMapper mapper,
+            AbstractValidator<DepartureDto> validator)
         {
             _repository = repository;
             _mapper = mapper;
+            _validator = validator;
         }
 
         public IEnumerable<DepartureDto> GetAll()
@@ -27,25 +34,62 @@ namespace Academy2018_.NET_Homework4.Core.Services
 
         public DepartureDto GetById(object id)
         {
-            return _mapper.Map<Departure, DepartureDto>(
+            var response = _mapper.Map<Departure, DepartureDto>(
                 _repository.Get().FirstOrDefault(d => d.Id == (int)id));
+
+            if (response == null)
+            {
+                throw new NotExistException();
+            }
+
+            return response;
         }
 
-        public void Add(DepartureDto dto)
+        public object Add(DepartureDto dto)
         {
-            _repository.Create(
-                _mapper.Map<DepartureDto, Departure>(dto));
+            var validationResult = _validator.Validate(dto);
+
+            if (validationResult.IsValid)
+            {
+                return _repository.Create(
+                    _mapper.Map<DepartureDto, Departure>(dto));
+            }
+
+            throw new ValidationException(validationResult.Errors);
         }
 
         public void Update(object id, DepartureDto dto)
         {
-            _repository.Update((int)id,
-                _mapper.Map<DepartureDto, Departure>(dto));
+            if (_repository.IsExist(id))
+            {
+                var validationResult = _validator.Validate(dto);
+
+                if (validationResult.IsValid)
+                {
+                    _repository.Update((int)id,
+                        _mapper.Map<DepartureDto, Departure>(dto));
+                }
+                else
+                {
+                    throw new ValidationException(validationResult.Errors);
+                }
+            }
+            else
+            {
+                throw new NotExistException();
+            }
         }
 
         public void Delete(object id)
         {
-            _repository.Delete((int)id);
+            if (_repository.IsExist(id))
+            {
+                _repository.Delete((int)id);
+            }
+            else
+            {
+                throw new NotExistException();
+            }
         }
     }
 }
