+++ Academy2018_.NET_Homework4.Core/Services/PilotsService.cs
@@ -1,10 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Security;
 using Academy2018_.NET_Homework4.Core.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Abstractions;
 using Academy2018_.NET_Homework4.Infrastructure.Models;
 using Academy2018_.NET_Homework4.Shared.DTOs;
+using Academy2018_.NET_Homework4.Shared.Exceptions;
 using AutoMapper;
+using FluentValidation;
+using FluentValidation.Results;
 
 namespace Academy2018_.NET_Homework4.Core.Services
 {
@@ -12,11 +16,16 @@ namespace Academy2018_.NET_Homework4.Core.Services
     {
         private readonly IRepository<Pilot> _repository;
         private readonly IMapper _mapper;
+        private readonly AbstractValidator<PilotDto> _validator;
 
-        public PilotsService(IRepository<Pilot> repository, IMapper mapper)
+        public PilotsService(
+            IRepository<Pilot> repository, 
+            IMapper mapper, 
+            AbstractValidator<PilotDto> validator)
         {
             _repository = repository;
             _mapper = mapper;
+            _validator = validator;
         }
 
         public IEnumerable<PilotDto> GetAll()
@@ -27,24 +36,62 @@ namespace Academy2018_.NET_Homework4.Core.Services
 
         public PilotDto GetById(object id)
         {
-            return _mapper.Map<Pilot, PilotDto>(
+            var response = _mapper.Map<Pilot, PilotDto>(
                 _repository.Get().FirstOrDefault(p => p.Id == (int)id));
+
+            if (response == null)
+            {
+                throw new NotExistException();
+            }
+
+            return response;
         }
 
-        public void Add(PilotDto pilotDto)
+        public object Add(PilotDto pilotDto)
         {
-            _repository.Create(
-                _mapper.Map<PilotDto, Pilot>(pilotDto));
+            var validationResult = _validator.Validate(pilotDto);
+
+            if (validationResult.IsValid)
+            {
+                return _repository.Create(
+                    _mapper.Map<PilotDto, Pilot>(pilotDto));
+            }
+
+            throw new ValidationException(validationResult.Errors);
         }
 
         public void Update(object id, PilotDto pilotDto)
         {
-            _repository.Update((int)id, _mapper.Map<PilotDto, Pilot>(pilotDto));
+            if (_repository.IsExist(id))
+            {
+                var validationResult = _validator.Validate(pilotDto);
+
+                if (validationResult.IsValid)
+                {
+                    _repository.Update((int)id,
+                        _mapper.Map<PilotDto, Pilot>(pilotDto));
+                }
+                else
+                {
+                    throw new ValidationException(validationResult.Errors);
+                }
+            }
+            else
+            {
+                throw new NotExistException();
+            }
         }
 
         public void Delete(object id)
         {
-            _repository.Delete((int)id);
+            if (_repository.IsExist(id))
+            {
+                _repository.Delete((int)id);
+            }
+            else
+            {
+                throw new NotExistException();
+            }
         }
     }
 }
